+++ tasmota/language/hu_HU.h
@@ -1356,4 +1356,7 @@
 #define D_MODBUSRELAY_RX         "MbsRelay RX"
 #define D_MODBUSRELAY_RX_ENA     "MbsRelay RX Ena"
 
+// xdrv_89_esp32_shelly_emulation.ino
+#define D_SHELLY                 "Shelly"
+
 #endif  // _LANGUAGE_HU_HU_H_