+++ tasmota/language/ro_RO.h
@@ -1349,4 +1349,7 @@
 #define D_MODBUSRELAY_RX         "MbsRelay RX"
 #define D_MODBUSRELAY_RX_ENA     "MbsRelay RX Ena"
 
+// xdrv_89_esp32_shelly_emulation.ino
+#define D_SHELLY                 "Shelly"
+
 #endif  // _LANGUAGE_RO_RO_H_