+++ tasmota/language/zh_TW.h
@@ -1351,4 +1351,7 @@
 #define D_MODBUSRELAY_RX         "MbsRelay RX"
 #define D_MODBUSRELAY_RX_ENA     "MbsRelay RX Ena"
 
+// xdrv_89_esp32_shelly_emulation.ino
+#define D_SHELLY                 "Shelly"
+
 #endif  // _LANGUAGE_ZH_TW_H_